Revert "CB-5332 fix multiple termination trigger"
This reverts commit 0b2f0a74764d6abd306fb2afabc0c10ac4711ac6.
