+++ core/src/main/java/com/sequenceiq/cloudbreak/controller/v4/UtilV4Controller.java
@@ -34,7 +34,7 @@ import com.sequenceiq.cloudbreak.service.account.PreferencesService;
 import com.sequenceiq.cloudbreak.service.cluster.RepositoryConfigValidationService;
 import com.sequenceiq.cloudbreak.service.filesystem.FileSystemSupportMatrixService;
 import com.sequenceiq.cloudbreak.service.securityrule.SecurityRuleService;
-import com.sequenceiq.cloudbreak.service.stack.flow.StackOperationService;
+import com.sequenceiq.cloudbreak.service.stack.StackService;
 import com.sequenceiq.cloudbreak.validation.externaldatabase.SupportedDatabaseProvider;
 import com.sequenceiq.common.api.util.versionchecker.ClientVersionUtil;
 import com.sequenceiq.common.api.util.versionchecker.VersionCheckResult;
@@ -71,7 +71,7 @@ public class UtilV4Controller extends NotificationController implements UtilV4En
     private UmsAuthorizationService umsAuthorizationService;
 
     @Inject
-    private StackOperationService stackOperationService;
+    private StackService stackService;
 
     @Value("${info.app.version:}")
     private String cbVersion;
@@ -133,7 +133,7 @@ public class UtilV4Controller extends NotificationController implements UtilV4En
 
     @Override
     public Response renewCertificate(RenewCertificateV4Request renewCertificateV4Request) {
-        stackOperationService.renewCertificate(renewCertificateV4Request.getStackName());
+        stackService.renewCertificate(renewCertificateV4Request.getStackName());
         return Response.ok().build();
     }
 }
