+++ core/src/main/java/com/sequenceiq/cloudbreak/job/StackStatusCheckerJob.java
@@ -47,7 +47,6 @@ import com.sequenceiq.cloudbreak.domain.stack.instance.InstanceMetaData;
 import com.sequenceiq.cloudbreak.logger.MDCBuilder;
 import com.sequenceiq.cloudbreak.service.cluster.ClusterApiConnectors;
 import com.sequenceiq.cloudbreak.service.cluster.ClusterService;
-import com.sequenceiq.cloudbreak.service.cluster.flow.ClusterOperationService;
 import com.sequenceiq.cloudbreak.service.environment.credential.CredentialConverter;
 import com.sequenceiq.cloudbreak.service.stack.InstanceMetaDataService;
 import com.sequenceiq.cloudbreak.service.stack.StackService;
@@ -71,9 +70,6 @@ public class StackStatusCheckerJob extends StatusCheckerJob {
 
     @Inject
     private ClusterService clusterService;
-
-    @Inject
-    private ClusterOperationService clusterOperationService;
 
     @Inject
     private ClusterApiConnectors clusterApiConnectors;
@@ -148,7 +144,7 @@ public class StackStatusCheckerJob extends StatusCheckerJob {
                 .filter(i -> !Set.of(SERVICES_UNHEALTHY, STOPPED).contains(i.getInstanceStatus()))
                 .map(InstanceMetaData::getDiscoveryFQDN)
                 .collect(toSet());
-        clusterOperationService.reportHealthChange(stack.getResourceCrn(), newFailedNodeNames, newHealtyHostNames);
+        clusterService.reportHealthChange(stack.getResourceCrn(), newFailedNodeNames, newHealtyHostNames);
         if (failedInstances.size() > 0) {
             clusterService.updateClusterStatusByStackId(stack.getId(), Status.AMBIGUOUS);
         } else if (EnumSet.of(Status.AMBIGUOUS, Status.STOPPED).contains(stack.getCluster().getStatus())) {
