+++ core/src/main/java/com/sequenceiq/cloudbreak/service/stack/StackService.java
@@ -1,6 +1,12 @@
 package com.sequenceiq.cloudbreak.service.stack;
 
+import static com.sequenceiq.cloudbreak.api.endpoint.v4.common.Status.AVAILABLE;
 import static com.sequenceiq.cloudbreak.api.endpoint.v4.common.Status.DELETE_IN_PROGRESS;
+import static com.sequenceiq.cloudbreak.api.endpoint.v4.common.Status.STOPPED;
+import static com.sequenceiq.cloudbreak.api.endpoint.v4.common.Status.STOP_REQUESTED;
+import static com.sequenceiq.cloudbreak.event.ResourceEvent.STACK_START_IGNORED;
+import static com.sequenceiq.cloudbreak.event.ResourceEvent.STACK_STOP_IGNORED;
+import static com.sequenceiq.cloudbreak.event.ResourceEvent.STACK_STOP_REQUESTED;
 import static com.sequenceiq.cloudbreak.exception.NotFoundException.notFound;
 import static com.sequenceiq.cloudbreak.util.Benchmark.measure;
 import static java.lang.String.format;
@@ -9,6 +15,8 @@ import java.io.IOException;
 import java.time.Duration;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -25,11 +33,15 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Strings;
+import com.sequenceiq.cloudbreak.api.endpoint.v4.autoscales.request.InstanceGroupAdjustmentV4Request;
+import com.sequenceiq.cloudbreak.api.endpoint.v4.common.DetailedStackStatus;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.common.StackType;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.common.Status;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.dto.NameOrCrn;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.base.InstanceStatus;
+import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.base.StatusRequest;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.request.StackV4Request;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.response.AutoscaleStackV4Response;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.response.StackV4Response;
@@ -54,8 +66,10 @@ import com.sequenceiq.cloudbreak.converter.stack.StackIdViewToStackResponseConve
 import com.sequenceiq.cloudbreak.core.CloudbreakImageCatalogException;
 import com.sequenceiq.cloudbreak.core.CloudbreakImageNotFoundException;
 import com.sequenceiq.cloudbreak.core.bootstrap.service.container.ContainerOrchestratorResolver;
+import com.sequenceiq.cloudbreak.core.flow2.service.ReactorFlowManager;
 import com.sequenceiq.cloudbreak.domain.Network;
 import com.sequenceiq.cloudbreak.domain.Orchestrator;
+import com.sequenceiq.cloudbreak.domain.StopRestrictionReason;
 import com.sequenceiq.cloudbreak.domain.projection.AutoscaleStack;
 import com.sequenceiq.cloudbreak.domain.projection.StackIdView;
 import com.sequenceiq.cloudbreak.domain.projection.StackListItem;
@@ -65,7 +79,10 @@ import com.sequenceiq.cloudbreak.domain.stack.Component;
 import com.sequenceiq.cloudbreak.domain.stack.Stack;
 import com.sequenceiq.cloudbreak.domain.stack.StackStatus;
 import com.sequenceiq.cloudbreak.domain.stack.StackValidation;
+import com.sequenceiq.cloudbreak.domain.stack.cluster.Cluster;
 import com.sequenceiq.cloudbreak.domain.stack.cluster.DatalakeResources;
+import com.sequenceiq.cloudbreak.domain.stack.cluster.host.HostGroup;
+import com.sequenceiq.cloudbreak.domain.stack.instance.InstanceGroup;
 import com.sequenceiq.cloudbreak.domain.stack.instance.InstanceMetaData;
 import com.sequenceiq.cloudbreak.domain.view.StackView;
 import com.sequenceiq.cloudbreak.exception.BadRequestException;
@@ -78,6 +95,8 @@ import com.sequenceiq.cloudbreak.orchestrator.model.OrchestrationCredential;
 import com.sequenceiq.cloudbreak.repository.StackRepository;
 import com.sequenceiq.cloudbreak.service.CloudbreakException;
 import com.sequenceiq.cloudbreak.service.ComponentConfigProviderService;
+import com.sequenceiq.cloudbreak.service.StackUpdater;
+import com.sequenceiq.cloudbreak.service.cluster.ClusterService;
 import com.sequenceiq.cloudbreak.service.datalake.DatalakeResourcesService;
 import com.sequenceiq.cloudbreak.service.decorator.StackResponseDecorator;
 import com.sequenceiq.cloudbreak.service.environment.credential.OpenSshPublicKeyValidator;
@@ -89,6 +108,7 @@ import com.sequenceiq.cloudbreak.service.stack.ShowTerminatedClusterConfigServic
 import com.sequenceiq.cloudbreak.service.stack.connector.adapter.ServiceProviderConnectorAdapter;
 import com.sequenceiq.cloudbreak.service.stackstatus.StackStatusService;
 import com.sequenceiq.cloudbreak.service.workspace.WorkspaceService;
+import com.sequenceiq.cloudbreak.structuredevent.event.CloudbreakEventService;
 import com.sequenceiq.cloudbreak.tag.AccountTagValidationFailed;
 import com.sequenceiq.cloudbreak.tag.CostTagging;
 import com.sequenceiq.cloudbreak.tag.request.CDPTagGenerationRequest;
@@ -99,6 +119,7 @@ import com.sequenceiq.cloudbreak.workspace.model.User;
 import com.sequenceiq.cloudbreak.workspace.model.Workspace;
 import com.sequenceiq.common.api.telemetry.model.Telemetry;
 import com.sequenceiq.flow.core.ResourceIdProvider;
+import com.sequenceiq.flow.api.model.FlowIdentifier;
 
 @Service
 public class StackService implements ResourceIdProvider {
@@ -130,6 +151,9 @@ public class StackService implements ResourceIdProvider {
     private ContainerOrchestratorResolver containerOrchestratorResolver;
 
     @Inject
+    private StackDownscaleValidatorService downscaleValidatorService;
+
+    @Inject
     private DatalakeResourcesService datalakeResourcesService;
 
     @Inject
@@ -157,6 +181,9 @@ public class StackService implements ResourceIdProvider {
     private TransactionService transactionService;
 
     @Inject
+    private CloudbreakEventService eventService;
+
+    @Inject
     private StackViewService stackViewService;
 
     @Inject
@@ -166,10 +193,19 @@ public class StackService implements ResourceIdProvider {
     private StackRepository stackRepository;
 
     @Inject
+    private ReactorFlowManager flowManager;
+
+    @Inject
+    private ClusterService clusterService;
+
+    @Inject
     private ConverterUtil converterUtil;
 
     @Inject
     private ImageService imageService;
+
+    @Inject
+    private StackUpdater stackUpdater;
 
     @Inject
     private CostTagging costTagging;
@@ -537,6 +573,42 @@ public class StackService implements ResourceIdProvider {
         stack.setPlatformVariant(connector.checkAndGetPlatformVariant(stack).value());
     }
 
+    public FlowIdentifier removeInstance(Stack stack, Long workspaceId, String instanceId, boolean forced, User user) {
+        InstanceMetaData metaData = validateInstanceForDownscale(instanceId, stack, workspaceId, user);
+        return flowManager.triggerStackRemoveInstance(stack.getId(), metaData.getInstanceGroupName(), metaData.getPrivateId(), forced);
+    }
+
+    public FlowIdentifier removeInstances(Stack stack, Long workspaceId, Collection<String> instanceIds, boolean forced, User user) {
+        Map<String, Set<Long>> instanceIdsByHostgroupMap = new HashMap<>();
+        for (String instanceId : instanceIds) {
+            InstanceMetaData metaData = validateInstanceForDownscale(instanceId, stack, workspaceId, user);
+            instanceIdsByHostgroupMap.computeIfAbsent(metaData.getInstanceGroupName(), s -> new LinkedHashSet<>()).add(metaData.getPrivateId());
+        }
+        return flowManager.triggerStackRemoveInstances(stack.getId(), instanceIdsByHostgroupMap, forced);
+    }
+
+    public FlowIdentifier updateStatus(Long stackId, StatusRequest status, boolean updateCluster, User user) {
+        Stack stack = getByIdWithLists(stackId);
+        Cluster cluster = null;
+        if (stack.getCluster() != null) {
+            cluster = clusterService.findOneWithLists(stack.getCluster().getId()).orElse(null);
+        }
+        switch (status) {
+            case SYNC:
+                return sync(stack, false, user);
+            case FULL_SYNC:
+                return sync(stack, true, user);
+            case REPAIR_FAILED_NODES:
+                return repairFailedNodes(stack, user);
+            case STOPPED:
+                return stop(stack, cluster, updateCluster, user);
+            case STARTED:
+                return start(stack, cluster, updateCluster, user);
+            default:
+                throw new BadRequestException("Cannot update the status of stack because status request not valid.");
+        }
+    }
+
     Optional<Stack> findTemplateWithLists(Long id) {
         return stackRepository.findTemplateWithLists(id);
     }
@@ -596,6 +668,10 @@ public class StackService implements ResourceIdProvider {
         return stackRepository.findByNetwork(network);
     }
 
+    public FlowIdentifier updateImage(Long stackId, Long workspaceId, String imageId, String imageCatalogName, String imageCatalogUrl, User user) {
+        return flowManager.triggerStackImageUpdate(stackId, imageId, imageCatalogName, imageCatalogUrl);
+    }
+
     @Override
     public Long getResourceIdByResourceCrn(String resourceCrn) {
         return getByCrn(resourceCrn).getId();
@@ -637,12 +713,145 @@ public class StackService implements ResourceIdProvider {
         return stackRepository.findByCrnAndWorkspaceIdWithLists(name, workspaceId, false, 0L);
     }
 
-    public Stack getByIdWithLists(Long id) {
+    private InstanceMetaData validateInstanceForDownscale(String instanceId, Stack stack, Long workspaceId, User user) {
+        InstanceMetaData metaData = instanceMetaDataService.findByStackIdAndInstanceId(stack.getId(), instanceId)
+                .orElseThrow(() -> new NotFoundException(format("Metadata for instance %s has not found.", instanceId)));
+        downscaleValidatorService.checkInstanceIsTheClusterManagerServerOrNot(metaData.getPublicIp(), metaData.getInstanceMetadataType());
+        downscaleValidatorService.checkClusterInValidStatus(stack.getCluster());
+        return metaData;
+    }
+
+    private Stack getByIdWithLists(Long id) {
         return stackRepository.findOneWithLists(id).orElseThrow(() -> new NotFoundException(format(STACK_NOT_FOUND_BY_ID_EXCEPTION_MESSAGE, id)));
     }
 
     private Stack getByCrnWithLists(String crn) {
         return stackRepository.findOneByCrnWithLists(crn).orElseThrow(NotFoundException.notFound("Stack", crn));
+    }
+
+    private FlowIdentifier repairFailedNodes(Stack stack, User user) {
+        LOGGER.debug("Received request to replace failed nodes: " + stack.getId());
+        return flowManager.triggerManualRepairFlow(stack.getId());
+    }
+
+    private FlowIdentifier sync(Stack stack, boolean full, User user) {
+        // TODO: is it a good condition?
+        if (!stack.isDeleteInProgress() && !stack.isStackInDeletionPhase() && !stack.isModificationInProgress()) {
+            if (full) {
+                return flowManager.triggerFullSync(stack.getId());
+            } else {
+                return flowManager.triggerStackSync(stack.getId());
+            }
+        } else {
+            LOGGER.debug("Stack could not be synchronized in {} state!", stack.getStatus());
+            return FlowIdentifier.notTriggered();
+        }
+    }
+
+    private FlowIdentifier stop(Stack stack, Cluster cluster, boolean updateCluster, User user) {
+        if (cluster != null && cluster.isStopInProgress()) {
+            setStackStatusToStopRequested(stack);
+            return FlowIdentifier.notTriggered();
+        } else {
+            return triggerStackStopIfNeeded(stack, cluster, updateCluster, user);
+        }
+    }
+
+    private FlowIdentifier triggerStackStopIfNeeded(Stack stack, Cluster cluster, boolean updateCluster, User user) {
+        if (!isStopNeeded(stack)) {
+            return FlowIdentifier.notTriggered();
+        }
+        if (cluster != null && !cluster.isStopped() && !stack.isStopFailed()) {
+            if (!updateCluster) {
+                throw new BadRequestException(format("Cannot update the status of stack '%s' to STOPPED, because the cluster's state is %s.",
+                        stack.getName(), cluster.getStatus().name()));
+            } else if (cluster.isClusterReadyForStop() || cluster.isStopFailed()) {
+                setStackStatusToStopRequested(stack);
+                return clusterService.updateStatus(stack.getId(), StatusRequest.STOPPED);
+            } else {
+                throw new BadRequestException(format("Cannot update the status of cluster '%s' to STOPPED, because the cluster's state is %s.",
+                        cluster.getName(), cluster.getStatus()));
+            }
+        } else {
+            stackUpdater.updateStackStatus(stack.getId(), DetailedStackStatus.STOP_REQUESTED);
+            return flowManager.triggerStackStop(stack.getId());
+        }
+    }
+
+    private boolean isStopNeeded(Stack stack) {
+        boolean result = true;
+        StopRestrictionReason reason = stack.isInfrastructureStoppable();
+        if (stack.isStopped()) {
+            eventService.fireCloudbreakEvent(stack.getId(), STOPPED.name(), STACK_STOP_IGNORED);
+            result = false;
+        } else if (reason != StopRestrictionReason.NONE) {
+            throw new BadRequestException(
+                    format("Cannot stop a stack '%s'. Reason: %s", stack.getName(), reason.getReason()));
+        } else if (!stack.isAvailable() && !stack.isStopFailed()) {
+            throw new BadRequestException(
+                    format("Cannot update the status of stack '%s' to STOPPED, because it isn't in AVAILABLE state.", stack.getName()));
+        }
+        return result;
+    }
+
+    private void setStackStatusToStopRequested(Stack stack) {
+        stackUpdater.updateStackStatus(stack.getId(), DetailedStackStatus.STOP_REQUESTED, "Stopping of cluster infrastructure has been requested.");
+        eventService.fireCloudbreakEvent(stack.getId(), STOP_REQUESTED.name(), STACK_STOP_REQUESTED);
+    }
+
+    @VisibleForTesting
+    FlowIdentifier start(Stack stack, Cluster cluster, boolean updateCluster, User user) {
+        FlowIdentifier flowIdentifier = FlowIdentifier.notTriggered();
+        if (stack.isAvailable() && (cluster == null || cluster.isAvailable())) {
+            eventService.fireCloudbreakEvent(stack.getId(), AVAILABLE.name(), STACK_START_IGNORED);
+        } else if (isStackStartable(stack) || isClusterStartable(cluster)) {
+            Stack startStack = stackUpdater.updateStackStatus(stack.getId(), DetailedStackStatus.START_REQUESTED);
+            flowIdentifier = flowManager.triggerStackStart(stack.getId());
+            if (updateCluster && cluster != null) {
+                clusterService.updateStatus(startStack, StatusRequest.STARTED);
+            }
+        } else {
+            throw new BadRequestException(format("Cannot update the status of stack '%s' to STARTED, because it is in %s state",
+                    stack.getName(), stack.getStatus().name()));
+        }
+        return flowIdentifier;
+    }
+
+    private boolean isClusterStartable(Cluster cluster) {
+        return cluster != null && (cluster.isStopped() || cluster.isStartFailed());
+    }
+
+    private boolean isStackStartable(Stack stack) {
+        return stack.isStopped() || stack.isStartFailed();
+    }
+
+    public FlowIdentifier updateNodeCount(Stack stack, InstanceGroupAdjustmentV4Request instanceGroupAdjustmentJson, boolean withClusterEvent, User user) {
+        try {
+            return transactionService.required(() -> {
+                Stack stackWithLists = getByIdWithLists(stack.getId());
+                validateStackStatus(stackWithLists);
+                validateInstanceGroup(stackWithLists, instanceGroupAdjustmentJson.getInstanceGroup());
+                validateScalingAdjustment(instanceGroupAdjustmentJson, stackWithLists);
+                validateInstanceStatuses(stackWithLists, instanceGroupAdjustmentJson);
+                if (withClusterEvent) {
+                    validateClusterStatus(stackWithLists);
+                    validateHostGroupAdjustment(instanceGroupAdjustmentJson, stackWithLists, instanceGroupAdjustmentJson.getScalingAdjustment());
+                    validataHostMetadataStatuses(stackWithLists, instanceGroupAdjustmentJson);
+                }
+                if (instanceGroupAdjustmentJson.getScalingAdjustment() > 0) {
+                    stackUpdater.updateStackStatus(stackWithLists.getId(), DetailedStackStatus.UPSCALE_REQUESTED);
+                    return flowManager.triggerStackUpscale(stackWithLists.getId(), instanceGroupAdjustmentJson, withClusterEvent);
+                } else {
+                    stackUpdater.updateStackStatus(stackWithLists.getId(), DetailedStackStatus.DOWNSCALE_REQUESTED);
+                    return flowManager.triggerStackDownscale(stackWithLists.getId(), instanceGroupAdjustmentJson);
+                }
+            });
+        } catch (TransactionExecutionException e) {
+            if (e.getCause() instanceof BadRequestException) {
+                throw e.getCause();
+            }
+            throw new TransactionRuntimeExecutionException(e);
+        }
     }
 
     public void updateMetaDataStatusIfFound(Long id, String hostName, InstanceStatus status, String statusReason) {
@@ -677,6 +886,88 @@ public class StackService implements ResourceIdProvider {
         return instanceMetaDataList.stream()
                 .filter(instanceMetaData -> hostNames.contains(instanceMetaData.getDiscoveryFQDN()))
                 .collect(Collectors.toSet());
+    }
+
+    private void validateScalingAdjustment(InstanceGroupAdjustmentV4Request instanceGroupAdjustmentJson, Stack stack) {
+        if (0 == instanceGroupAdjustmentJson.getScalingAdjustment()) {
+            throw new BadRequestException(format("Requested scaling adjustment on stack '%s' is 0. Nothing to do.", stack.getName()));
+        }
+        if (0 > instanceGroupAdjustmentJson.getScalingAdjustment()) {
+            InstanceGroup instanceGroup = stack.getInstanceGroupByInstanceGroupName(instanceGroupAdjustmentJson.getInstanceGroup());
+            if (-1 * instanceGroupAdjustmentJson.getScalingAdjustment() > instanceGroup.getNodeCount()) {
+                throw new BadRequestException(format("There are %s instances in instance group '%s'. Cannot remove %s instances.",
+                        instanceGroup.getNodeCount(), instanceGroup.getGroupName(),
+                        -1 * instanceGroupAdjustmentJson.getScalingAdjustment()));
+            }
+            int removableHosts = instanceMetaDataService.findRemovableInstances(stack.getId(), instanceGroupAdjustmentJson.getInstanceGroup()).size();
+            if (removableHosts < -1 * instanceGroupAdjustmentJson.getScalingAdjustment()) {
+                throw new BadRequestException(
+                        format("There are %s unregistered instances in instance group '%s' but %s were requested. Decommission nodes from the cluster!",
+                                removableHosts, instanceGroup.getGroupName(), instanceGroupAdjustmentJson.getScalingAdjustment() * -1));
+            }
+        }
+    }
+
+    private void validateInstanceStatuses(Stack stack, InstanceGroupAdjustmentV4Request instanceGroupAdjustmentJson) {
+        if (instanceGroupAdjustmentJson.getScalingAdjustment() > 0) {
+            List<InstanceMetaData> instanceMetaDataList =
+                    stack.getInstanceMetaDataAsList().stream().filter(im -> !im.isTerminated() && !im.isRunning() && !im.isCreated())
+                            .collect(Collectors.toList());
+            if (!instanceMetaDataList.isEmpty()) {
+                String ims = instanceMetaDataList.stream()
+                        .map(im -> im.getInstanceId() != null ? im.getInstanceId() : im.getPrivateId() + ": " + im.getInstanceStatus())
+                        .collect(Collectors.joining(", "));
+                throw new BadRequestException(
+                        format("Upscale is not allowed because the following instances are not in running state: %s. Please remove them first!", ims));
+            }
+        }
+    }
+
+    private void validataHostMetadataStatuses(Stack stack, InstanceGroupAdjustmentV4Request instanceGroupAdjustmentJson) {
+        if (instanceGroupAdjustmentJson.getScalingAdjustment() > 0) {
+            List<InstanceMetaData> instanceMetaDataAsList = stack.getInstanceMetaDataAsList();
+            List<InstanceMetaData> unhealthyInstanceMetadataList = instanceMetaDataAsList.stream()
+                    .filter(instanceMetaData -> InstanceStatus.SERVICES_UNHEALTHY.equals(instanceMetaData.getInstanceStatus()))
+                    .collect(Collectors.toList());
+            if (!unhealthyInstanceMetadataList.isEmpty()) {
+                String notHealthyInstances = unhealthyInstanceMetadataList.stream()
+                        .map(instanceMetaData -> instanceMetaData.getDiscoveryFQDN() + ": " + instanceMetaData.getInstanceStatus())
+                        .collect(Collectors.joining(","));
+                throw new BadRequestException(
+                        format("Upscale is not allowed because the following hosts are not healthy: %s. Please remove them first!", notHealthyInstances));
+            }
+        }
+    }
+
+    private void validateHostGroupAdjustment(InstanceGroupAdjustmentV4Request instanceGroupAdjustmentJson, Stack stack, Integer adjustment) {
+        Optional<HostGroup> hostGroup = stack.getCluster().getHostGroups().stream()
+                .filter(input -> input.getInstanceGroup().getGroupName().equals(instanceGroupAdjustmentJson.getInstanceGroup())).findFirst();
+        if (!hostGroup.isPresent()) {
+            throw new BadRequestException(format("Instancegroup '%s' not found or not part of stack '%s'",
+                    instanceGroupAdjustmentJson.getInstanceGroup(), stack.getName()));
+        }
+    }
+
+    private void validateStackStatus(Stack stack) {
+        if (!stack.isAvailable()) {
+            throw new BadRequestException(format("Stack '%s' is currently in '%s' state. Node count can only be updated if it's running.",
+                    stack.getName(), stack.getStatus()));
+        }
+    }
+
+    private void validateClusterStatus(Stack stack) {
+        Cluster cluster = stack.getCluster();
+        if (cluster != null && !cluster.isAvailable()) {
+            throw new BadRequestException(format("Cluster '%s' is currently in '%s' state. Node count can only be updated if it's not available.",
+                    cluster.getName(), cluster.getStatus()));
+        }
+    }
+
+    private void validateInstanceGroup(Stack stack, String instanceGroupName) {
+        InstanceGroup instanceGroup = stack.getInstanceGroupByInstanceGroupName(instanceGroupName);
+        if (instanceGroup == null) {
+            throw new BadRequestException(format("Stack '%s' does not have an instanceGroup named '%s'.", stack.getName(), instanceGroupName));
+        }
     }
 
     private void addTemplateForStack(Stack stack, String template) {
@@ -773,6 +1064,12 @@ public class StackService implements ResourceIdProvider {
 
     public int setMinaSshdServiceIdByStackId(Long id, String minaSshdServiceId) {
         return stackRepository.setMinaSshdServiceIdByStackId(id, minaSshdServiceId);
+    }
+
+    public void renewCertificate(String stackName) {
+        Workspace workspace = workspaceService.getForCurrentUser();
+        Stack stack = getByNameInWorkspace(stackName, workspace.getId());
+        flowManager.triggerClusterCertificationRenewal(stack.getId());
     }
 
     StackRepository repository() {
