+++ core/src/test/java/com/sequenceiq/cloudbreak/job/StackStatusCheckerJobTest.java
@@ -36,7 +36,6 @@ import com.sequenceiq.cloudbreak.domain.stack.cluster.Cluster;
 import com.sequenceiq.cloudbreak.domain.stack.instance.InstanceMetaData;
 import com.sequenceiq.cloudbreak.service.cluster.ClusterApiConnectors;
 import com.sequenceiq.cloudbreak.service.cluster.ClusterService;
-import com.sequenceiq.cloudbreak.service.cluster.flow.ClusterOperationService;
 import com.sequenceiq.cloudbreak.service.environment.credential.CredentialConverter;
 import com.sequenceiq.cloudbreak.service.stack.InstanceMetaDataService;
 import com.sequenceiq.cloudbreak.service.stack.StackService;
@@ -83,9 +82,6 @@ public class StackStatusCheckerJobTest {
 
     @Mock
     private ClusterService clusterService;
-
-    @Mock
-    private ClusterOperationService clusterOperationService;
 
     @Mock
     private ClusterApi clusterApi;
@@ -206,7 +202,7 @@ public class StackStatusCheckerJobTest {
         setupForCM();
         underTest.executeInternal(jobExecutionContext);
 
-        verify(clusterOperationService, times(0)).reportHealthChange(anyString(), anySet(), anySet());
+        verify(clusterService, times(0)).reportHealthChange(anyString(), anySet(), anySet());
         verify(instanceStateQuery, times(1)).getCloudVmInstanceStatuses(any(), any(), any());
     }
 
@@ -218,7 +214,7 @@ public class StackStatusCheckerJobTest {
         when(clusterApi.clusterStatusService()).thenReturn(clusterStatusService);
         underTest.executeInternal(jobExecutionContext);
 
-        verify(clusterOperationService, times(1)).reportHealthChange(any(), anySet(), anySet());
+        verify(clusterService, times(1)).reportHealthChange(any(), anySet(), anySet());
         verify(instanceStateQuery, times(1)).getCloudVmInstanceStatuses(any(), any(), any());
     }
 }
