+++ flow/src/main/java/com/sequenceiq/flow/repository/FlowLogRepository.java
@@ -48,8 +48,6 @@ public interface FlowLogRepository extends CrudRepository<FlowLog, Long> {
 
     List<FlowLog> findAllByResourceIdOrderByCreatedDesc(Long resourceId);
 
-    List<FlowLog> findAllByResourceIdAndFinalizedIsFalseOrderByCreatedDesc(Long resourceId);
-
     List<FlowLog> findAllByFlowIdOrderByCreatedDesc(String flowId);
 
     List<FlowLog> findAllByResourceIdOrderByCreatedDesc(Long resourceId, Pageable page);
